SH-199 Move config folder "up" and update includes
