+++ modules/config/common/basic_esp32_yd_config.h
@@ -1,8 +1,8 @@
 #pragma once
 #ifdef HC12_MODULE
-    #include "config/hc12_common_config.h"
+    #include "hc12_common_config.h"
 
-    #define BAUD_RATE 9600
+#define BAUD_RATE 9600
     #define RX_PIN 6
     #define TX_PIN 7
     #define SET_PIN 5
@@ -11,4 +11,5 @@
 
 #define LED_PIN 39
 #define BUTTON_PIN 1
+
 #define CENTRAL_UNIT
