+++ modules/config/common/hc12_common_config.h
