+++ modules/platformio.ini
@@ -14,6 +14,8 @@ platform = espressif32
 board = esp32dev
 upload_speed = 921600
 framework = arduino
+lib_deps =
+    johboh/nlohmann-json@^3.12.0
 monitor_filters =
     esp32_exception_decoder
 build_flags =
@@ -27,6 +29,8 @@ build_flags =
 ;board = esp32-s3-devkitm-1
 ;upload_speed = 921600
 ;framework = arduino
+;lib_deps =
+;    johboh/nlohmann-json@^3.12.0
 ;monitor_filters =
 ;    esp32_exception_decoder
 ;build_flags =
@@ -40,6 +44,8 @@ platform = espressif32
 board = esp32-s3-devkitm-1
 upload_speed = 921600
 framework = arduino
+lib_deps =
+    johboh/nlohmann-json@^3.12.0
 monitor_filters =
     esp32_exception_decoder
 build_flags =
