+++ modules/src/communication/addressing/central_unit_addressing.cpp
@@ -1,14 +1,6 @@
 #include "central_unit_addressing.h"
 
-#include <Arduino.h>
-#include <memory>
-
-#include "smart_home_config.h"
-#include "config/communication_config.h"
-#include "config/addressing_config.h"
-
-#include "../../utils/uint8_array_handlers.h"
-#include "utils/logger.h"
+#include "utils/uint8_array_handlers.h"
 #include "communication/communication.h"
 
 namespace uah = Utils::ArrayHandlers;
