+++ modules/src/communication/addressing/module_addressing.cpp
@@ -1,10 +1,6 @@
 #include "module_addressing.h"
 
-#include "smart_home_config.h"
-#include "config/communication_config.h"
-
-#include "../../utils/uint8_array_handlers.h"
-#include "../../utils/logger.h"
+#include "utils/uint8_array_handlers.h"
 
 #include "communication/communication.h"
