+++ modules/src/universal_module_system/data_manager.cpp
@@ -1,0 +1,8 @@
+#include "data_manager.h"
+
+#include "../config/universal_module_system_config.h"
+
+namespace UniversalModuleSystem {
+    // TODO !BEFORE PULL REQUEST! remove
+    // SPIFFS.begin(true);
+}
