+++ modules/src/universal_module_system/data_manager.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include <Arduino.h>
+#include <memory>
+#include <nlohmann/json.hpp>
+#include <SPIFFS.h>
+
+#include "utils/logger.h"
+
+namespace ul = Utils::Logging;
+
+namespace UniversalModuleSystem {
+    class DataManager {
+        // jso
+    };
+}
