+++ modules/src/utils/uint8_array_handlers.h
@@ -1,4 +1,5 @@
 #pragma once
+
 #include <Arduino.h>
 
 namespace Utils {
